Refactor generation API into modular routers
