+++ backend/api/v1/__init__.py
@@ -3,4 +3,4 @@
 This package exposes submodules for adapters, compose, deliveries, and recommendations.
 """
 
-from . import adapters, analytics, compose, deliveries, recommendations
+from . import adapters, analytics, compose, deliveries, generation, recommendations
