--- backend/api/v1/generation.py
@@ -1,403 +1,0 @@
-"""Router for SDNext generation endpoints."""
-
-from typing import Dict, List
-
-import structlog
-from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query, Response
-from fastapi.responses import StreamingResponse
-
-from backend.core.dependencies import (
-    get_application_services,
-    get_core_services,
-    get_domain_services,
-)
-from backend.delivery import get_generation_backend
-from backend.schemas import (
-    ComposeRequest,
-    DeliveryCreateResponse,
-    DeliveryRead,
-    DeliveryWrapper,
-    GenerationBulkDeleteRequest,
-    GenerationCancelResponse,
-    GenerationExportRequest,
-    GenerationJobStatus,
-    GenerationMode,
-    GenerationResultFormat,
-    GenerationResultSummary,
-    SDNextGenerationParams,
-    SDNextGenerationResult,
-)
-from backend.services import ApplicationServices, CoreServices, DomainServices
-from backend.services.generation import normalize_generation_status
-from backend.services.generation.presenter import build_active_job, build_result
-
-ACTIVE_JOB_STATUSES = {"pending", "running"}
-CANCELLABLE_STATUSES = {"pending", "running"}
-
-logger = structlog.get_logger(__name__)
-
-router = APIRouter(prefix="/generation", tags=["generation"])
-
-
-async def _execute_generation_request(
-    *,
-    generation_params: SDNextGenerationParams,
-    backend: str,
-    mode: GenerationMode,
-    save_images: bool,
-    return_format: GenerationResultFormat,
-    domain: DomainServices,
-    application: ApplicationServices,
-) -> SDNextGenerationResult:
-    """Validate and orchestrate a generation request."""
-    warnings = await domain.generation.validate_generation_params(generation_params)
-    if warnings:
-        logger.warning(
-            "generation parameter validation warnings",
-            warnings=warnings,
-            backend=backend,
-            mode=mode.value,
-        )
-
-    params = {
-        "mode": mode.value,
-        "save_images": save_images,
-        "return_format": return_format.value,
-    }
-
-    result = await domain.generation.generate_image(
-        generation_params.prompt,
-        backend,
-        generation_params,
-        **params,
-    )
-
-    if result.job_id and mode == GenerationMode.DEFERRED:
-        await application.generation_coordinator.broadcast_job_started(
-            result.job_id,
-            generation_params,
-        )
-
-    return result
-
-
-@router.get("/backends", response_model=Dict[str, bool])
-async def list_generation_backends():
-    """List available generation backends and their status."""
-    from backend.delivery.base import delivery_registry
-
-    backends_info = delivery_registry.list_available_backends()
-    return backends_info.get("generation", {})
-
-
-@router.post("/generate", response_model=SDNextGenerationResult)
-async def generate_image(
-    generation_params: SDNextGenerationParams,
-    backend: str = "sdnext",
-    mode: GenerationMode = GenerationMode.IMMEDIATE,
-    save_images: bool = True,
-    return_format: GenerationResultFormat = GenerationResultFormat.BASE64,
-    domain: DomainServices = Depends(get_domain_services),
-    application: ApplicationServices = Depends(get_application_services),
-):
-    """Generate an image using the specified backend.
-
-    Args:
-        generation_params: Generation parameters
-        backend: Generation backend name
-        mode: Generation mode ("immediate" or "deferred")
-        save_images: Whether to save generated images
-        return_format: Return format ("base64", "file_path", "url")
-        domain: Domain service container providing adapter access
-        application: Application service container orchestrating deliveries
-
-    """
-    return await _execute_generation_request(
-        generation_params=generation_params,
-        backend=backend,
-        mode=mode,
-        save_images=save_images,
-        return_format=return_format,
-        domain=domain,
-        application=application,
-    )
-
-
-@router.get("/progress/{job_id}", response_model=SDNextGenerationResult)
-async def check_generation_progress(
-    job_id: str,
-    backend: str = "sdnext",
-):
-    """Check the progress of a generation job."""
-    try:
-        generation_backend = get_generation_backend(backend)
-    except ValueError as exc:
-        raise HTTPException(
-            status_code=404, detail=f"Backend '{backend}' not found"
-        ) from exc
-
-    result = await generation_backend.check_progress(job_id)
-    normalized_status = normalize_generation_status(result.status)
-    if normalized_status != result.status:
-        result = result.model_copy(update={"status": normalized_status})
-    return result
-
-
-@router.post("/compose-and-generate", response_model=SDNextGenerationResult)
-async def compose_and_generate(
-    compose_request: ComposeRequest,
-    generation_params: SDNextGenerationParams,
-    backend: str = "sdnext",
-    mode: GenerationMode = GenerationMode.IMMEDIATE,
-    save_images: bool = True,
-    return_format: GenerationResultFormat = GenerationResultFormat.BASE64,
-    domain: DomainServices = Depends(get_domain_services),
-    application: ApplicationServices = Depends(get_application_services),
-):
-    """Compose LoRA prompt and immediately generate images.
-
-    This endpoint combines the compose and generate operations for convenience.
-    """
-    # Step 1: Compose prompt via shared helper
-    composition = domain.compose.compose_from_adapter_service(
-        domain.adapters,
-        prefix=compose_request.prefix or "",
-        suffix=compose_request.suffix or "",
-    )
-
-    if not composition.tokens:
-        raise HTTPException(status_code=400, detail="No active adapters found")
-
-    if composition.warnings:
-        logger.warning(
-            "composition warnings during compose-and-generate",
-            warnings=composition.warnings,
-            backend=backend,
-            mode=mode.value,
-        )
-
-    composed_params = generation_params.model_copy(
-        update={"prompt": composition.prompt}
-    )
-
-    return await _execute_generation_request(
-        generation_params=composed_params,
-        backend=backend,
-        mode=mode,
-        save_images=save_images,
-        return_format=return_format,
-        domain=domain,
-        application=application,
-    )
-
-
-@router.post("/queue-generation", response_model=DeliveryCreateResponse)
-async def queue_generation_job(
-    generation_params: SDNextGenerationParams,
-    background_tasks: BackgroundTasks,
-    backend: str = "sdnext",
-    mode: str = "deferred",
-    save_images: bool = True,
-    return_format: str = "base64",
-    domain: DomainServices = Depends(get_domain_services),
-    application: ApplicationServices = Depends(get_application_services),
-):
-    """Queue a generation job for background processing.
-
-    This creates a delivery job that will be processed by a worker.
-    """
-    coordinator = application.generation_coordinator
-
-    job = coordinator.schedule_generation_job(
-        generation_params,
-        backend=backend,
-        mode=mode,
-        save_images=save_images,
-        return_format=return_format,
-        background_tasks=background_tasks,
-    )
-
-    await coordinator.broadcast_job_started(job.id, generation_params)
-
-    return DeliveryCreateResponse(delivery_id=job.id)
-
-
-@router.get("/jobs/active", response_model=List[GenerationJobStatus])
-async def list_active_generation_jobs(
-    limit: int = Query(50, ge=1, le=200),
-    services: ApplicationServices = Depends(get_application_services),
-):
-    """Return active generation jobs for frontend queues."""
-    coordinator = services.generation_coordinator
-    jobs_by_id = {}
-    for status in ACTIVE_JOB_STATUSES:
-        for job in services.deliveries.list_jobs(status=status, limit=limit):
-            jobs_by_id[job.id] = job
-
-    ordered_jobs = sorted(
-        jobs_by_id.values(),
-        key=lambda job: job.started_at or job.created_at,
-        reverse=True,
-    )
-
-    return [build_active_job(job, coordinator) for job in ordered_jobs[:limit]]
-
-
-@router.get("/jobs/{job_id}", response_model=DeliveryWrapper)
-async def get_generation_job(
-    job_id: str,
-    services: ApplicationServices = Depends(get_application_services),
-):
-    """Get generation job status and results."""
-    job = services.deliveries.get_job(job_id)
-    if not job:
-        raise HTTPException(status_code=404, detail="Job not found")
-
-    presenter = build_active_job(job, services.generation_coordinator)
-
-    params = services.deliveries.get_job_params(job)
-    if not isinstance(params, dict):
-        params = {}
-
-    delivery_read = DeliveryRead(
-        id=job.id,
-        prompt=job.prompt,
-        mode=job.mode,
-        params=params,
-        status=presenter.status,
-        result=presenter.result,
-        created_at=job.created_at,
-        started_at=job.started_at,
-        finished_at=job.finished_at,
-    )
-
-    return DeliveryWrapper(delivery=delivery_read)
-
-
-@router.post("/jobs/{job_id}/cancel", response_model=GenerationCancelResponse)
-async def cancel_generation_job(
-    job_id: str,
-    services: ApplicationServices = Depends(get_application_services),
-):
-    """Cancel an active generation job."""
-    job = services.deliveries.get_job(job_id)
-    if job is None:
-        raise HTTPException(status_code=404, detail="Job not found")
-
-    if job.status not in CANCELLABLE_STATUSES:
-        raise HTTPException(
-            status_code=400, detail="Job cannot be cancelled in its current state"
-        )
-
-    services.deliveries.update_job_status(
-        job_id, "cancelled", result={"status": "cancelled"}
-    )
-    services.websocket.stop_job_monitoring(job_id)
-
-    return GenerationCancelResponse(
-        status="cancelled", message="Generation job cancelled"
-    )
-
-
-@router.get("/results", response_model=List[GenerationResultSummary])
-async def list_generation_results(
-    limit: int = Query(20, ge=1, le=100),
-    offset: int = Query(0, ge=0),
-    services: ApplicationServices = Depends(get_application_services),
-):
-    """Return recent completed generation results."""
-    jobs = services.deliveries.list_jobs(status="succeeded", limit=limit, offset=offset)
-
-    coordinator = services.generation_coordinator
-    return [build_result(job, coordinator) for job in jobs]
-
-
-@router.delete("/results/bulk-delete")
-async def bulk_delete_generation_results(
-    request: GenerationBulkDeleteRequest,
-    application: ApplicationServices = Depends(get_application_services),
-    core: CoreServices = Depends(get_core_services),
-):
-    """Delete multiple generation results in a single request."""
-    target_ids = [str(identifier) for identifier in request.ids]
-    deleted = application.deliveries.bulk_delete_job_results(
-        target_ids,
-        storage=core.storage,
-        coordinator=application.generation_coordinator,
-    )
-    return {"deleted": deleted}
-
-
-@router.delete("/results/{result_id}", status_code=204)
-async def delete_generation_result(
-    result_id: str,
-    application: ApplicationServices = Depends(get_application_services),
-    core: CoreServices = Depends(get_core_services),
-):
-    """Delete a stored generation result and its associated artifacts."""
-    deleted = application.deliveries.delete_job_result(
-        str(result_id),
-        storage=core.storage,
-        coordinator=application.generation_coordinator,
-    )
-    if not deleted:
-        raise HTTPException(status_code=404, detail="Result not found")
-
-    return Response(status_code=204)
-
-
-@router.post("/results/export")
-async def export_generation_results(
-    request: GenerationExportRequest,
-    application: ApplicationServices = Depends(get_application_services),
-    core: CoreServices = Depends(get_core_services),
-):
-    """Stream a ZIP archive containing the requested generation results."""
-    target_ids = [str(identifier) for identifier in request.ids]
-    archive = application.deliveries.build_results_archive(
-        target_ids,
-        storage=core.storage,
-        coordinator=application.generation_coordinator,
-        include_metadata=request.include_metadata,
-    )
-    if archive is None:
-        raise HTTPException(status_code=404, detail="No results found")
-
-    headers = {
-        "Content-Disposition": f'attachment; filename="{archive.filename}"',
-        "Content-Length": str(archive.size),
-    }
-
-    return StreamingResponse(
-        archive.iterator, media_type="application/zip", headers=headers
-    )
-
-
-@router.get("/results/{result_id}/download")
-async def download_generation_result(
-    result_id: str,
-    application: ApplicationServices = Depends(get_application_services),
-    core: CoreServices = Depends(get_core_services),
-):
-    """Download the primary asset associated with a generation result."""
-    job = application.deliveries.get_job(result_id)
-    if job is None:
-        raise HTTPException(status_code=404, detail="Result not found")
-
-    payload = application.deliveries.build_result_download(
-        job,
-        storage=core.storage,
-        coordinator=application.generation_coordinator,
-    )
-    if payload is None:
-        raise HTTPException(status_code=404, detail="Result asset is unavailable")
-
-    headers = {
-        "Content-Disposition": f'attachment; filename="{payload.filename}"',
-    }
-    if payload.size is not None:
-        headers["Content-Length"] = str(payload.size)
-
-    return StreamingResponse(
-        payload.iterator, media_type=payload.content_type, headers=headers
-    )
