+++ backend/api/v1/generation/controllers/__init__.py
@@ -1,0 +1,5 @@
+"""Controller helpers for generation API groups."""
+
+from . import jobs, live, results
+
+__all__ = ["jobs", "live", "results"]
