+++ backend/api/v1/generation/presenters/__init__.py
@@ -1,0 +1,12 @@
+"""Presenter utilities for generation API responses."""
+
+from .jobs import build_active_job
+from .results import build_result
+from .streams import stream_archive, stream_download
+
+__all__ = [
+    "build_active_job",
+    "build_result",
+    "stream_archive",
+    "stream_download",
+]
