+++ backend/api/v1/generation/presenters/results.py
@@ -1,4 +1,4 @@
-"""Presenter helpers for generation delivery jobs."""
+"""Presentation helpers for generation results."""
 
 from __future__ import annotations
 
@@ -6,9 +6,10 @@ import json
 from typing import Any, Dict, Optional
 
 from backend.models import DeliveryJob
-from backend.schemas import GenerationJobStatus, GenerationResultSummary
+from backend.schemas import GenerationResultSummary
+from backend.services.generation import GenerationCoordinator, normalize_generation_status
 
-from . import GenerationCoordinator, normalize_generation_status
+__all__ = ["build_result"]
 
 
 def _extract_image_url(result_payload: Dict[str, Any]) -> Optional[str]:
@@ -33,31 +34,6 @@ def _coerce_generation_info(result_payload: Dict[str, Any]) -> Optional[Dict[str
         return parsed if isinstance(parsed, dict) else None
 
     return None
-
-
-def build_active_job(
-    job: DeliveryJob,
-    coordinator: GenerationCoordinator,
-) -> GenerationJobStatus:
-    """Return an API-facing representation of an active job."""
-    serialized = coordinator.serialize_delivery_job(job)
-    params = serialized.get("params") or {}
-    result_payload = serialized.get("result") or {}
-
-    return GenerationJobStatus(
-        id=job.id,
-        jobId=job.id,
-        prompt=params.get("prompt") or job.prompt,
-        status=normalize_generation_status(job.status),
-        progress=serialized.get("progress", 0.0),
-        message=serialized.get("message"),
-        error=serialized.get("error"),
-        params=params,
-        created_at=job.created_at,
-        startTime=job.started_at or job.created_at,
-        finished_at=job.finished_at,
-        result=result_payload or None,
-    )
 
 
 def build_result(
